requirements.txt
flask
python-dotenv
google-generativeai
flask-cors
Pillow
pymongo